refactor: mapping ErrorNotifier is in Mapping module now
