+++ spec/cocina/models/mapping/from_mods/descriptive/subject_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe Cocina::Models::Mapping::FromMods::Descriptive::Subject do
 
   let(:descriptive_builder) { Cocina::Models::Mapping::FromMods::Descriptive::DescriptiveBuilder.new(notifier: notifier) }
 
-  let(:notifier) { instance_double(Cocina::Models::Mapping::FromMods::ErrorNotifier) }
+  let(:notifier) { instance_double(Cocina::Models::Mapping::ErrorNotifier) }
 
   let(:ng_xml) do
     Nokogiri::XML <<~XML
