+++ spec/cocina/models/mapping/from_mods/descriptive/titles_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe Cocina::Models::Mapping::FromMods::Descriptive::Titles do
       described_class.build(resource_element: ng_xml.root, require_title: require_title, notifier: notifier)
     end
 
-    let(:notifier) { instance_double(Cocina::Models::Mapping::FromMods::ErrorNotifier) }
+    let(:notifier) { instance_double(Cocina::Models::Mapping::ErrorNotifier) }
 
     let(:require_title) { true }
